Fixed controller with my class Order instead Sort.Direction
