+++ review-service/src/main/java/com/java/akdev/reviewservice/controller/ReviewController.java
@@ -3,13 +3,13 @@ package com.java.akdev.reviewservice.controller;
 import com.java.akdev.reviewservice.dto.ReviewCreateDto;
 import com.java.akdev.reviewservice.dto.ReviewReadDto;
 import com.java.akdev.reviewservice.dto.ReviewResponse;
+import com.java.akdev.reviewservice.enumeration.Order;
 import com.java.akdev.reviewservice.enumeration.Receiver;
 import com.java.akdev.reviewservice.enumeration.SortField;
 import com.java.akdev.reviewservice.service.ReviewService;
 import jakarta.validation.constraints.Min;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Sort;
 import org.springframework.http.ResponseEntity;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
@@ -27,7 +27,7 @@ public class ReviewController {
     public ResponseEntity<Page<ReviewReadDto>> findAll(@RequestParam @Min(1) Integer page,
                                                        @RequestParam @Min(1) Integer size,
                                                        @RequestParam SortField sortField,
-                                                       @RequestParam Sort.Direction order) {
+                                                       @RequestParam Order order) {
         return ResponseEntity.status(200)
                 .body(reviewService.findAll(page, size, sortField, order));
     }
