+++ review-service/src/main/java/com/java/akdev/reviewservice/enumeration/Order.java
@@ -1,0 +1,5 @@
+package com.java.akdev.reviewservice.enumeration;
+
+public enum Order {
+    ASC, DESC
+}
